* bfd-in.h: Comment typo fix.  Formatting. 	* bfd-in2.h: Regenerate. 	* coff64-rs6000.c (xcoff64_openr_next_archived_file): Warning fix. 	* elf32-m68hc12.c (m68hc12_elf_set_mach_from_flags): Prototype. 	* elf64-mmix.c (mmix_dump_bpo_gregs): Warning fix.

+++ bfd/elf64-mmix.c
@@ -2281,8 +2281,8 @@ mmix_dump_bpo_gregs (link_info, pf)
     for (i = 0; i < gregdata->n_max_bpo_relocs; i++)
       (*pf) ("%4u (%4u)/%4u#%u: 0x%08lx%08lx  r: %3u o: %3u\n",
 	     i,
-	     gregdata->bpo_reloc_indexes != NULL
-	     ? gregdata->bpo_reloc_indexes[i] : -1,
+	     (gregdata->bpo_reloc_indexes != NULL
+	      ? gregdata->bpo_reloc_indexes[i] : (size_t) -1),
 	     gregdata->reloc_request[i].bpo_reloc_no,
 	     gregdata->reloc_request[i].valid,
